[TWINFACES-568] feat: form refactor resolve PR comments
